+++ lib/app_state.dart
@@ -1,6 +1,3 @@
-export 'src/blocs/app/app.dart'                               show AppBloc;
-export 'src/blocs/app/configuration.dart'                     show AppConfiguration;
-
 export 'src/blocs/page/configuration.dart'                    show PageConfiguration;
 export 'src/blocs/page/close_event.dart'                      show PageBlocCloseEvent;
 export 'src/blocs/page/configuration_changed_event.dart'      show PageBlocConfigurationChangedEvent;
@@ -14,8 +11,16 @@ export 'src/blocs/page_stack/duplicate_page_key_action.dart'  show DuplicatePage
 export 'src/blocs/page_stack/event.dart'                      show PageStackBlocEvent;
 export 'src/blocs/page_stack/page_stack.dart'                 show PageStackBloc;
 export 'src/blocs/page_stack/page_event.dart'                 show PageStackPageBlocEvent;
+export 'src/blocs/page_stack/route_information_parser.dart'   show PageStackRouteInformationParser;
+export 'src/blocs/page_stack/router_delegate.dart'            show PageStackRouterDelegate;
 
-export 'src/delegates/page_stack.dart'                        show PageStackRouterDelegate;
+export 'src/blocs/page_stacks/back_button_dispatcher.dart'            show PageStacksBackButtonDispatcher;
+export 'src/blocs/page_stacks/configuration.dart'                     show PageStacksConfiguration;
+export 'src/blocs/page_stacks/current_page_stack_changed_event.dart'  show CurrentPageStackChangedEvent;
+export 'src/blocs/page_stacks/event.dart'                             show PageStacksBlocEvent;
+export 'src/blocs/page_stacks/page_stacks.dart'                       show PageStacksBloc;
+export 'src/blocs/page_stacks/route_information_parser.dart'          show PageStacksRouteInformationParser;
+export 'src/blocs/page_stacks/router_delegate.dart'                   show PageStacksRouterDelegate;
 
 export 'src/models/back_pressed_result_enum.dart'             show BackPressedResult;
 
@@ -23,7 +28,5 @@ export 'src/pages/abstract.dart'                              show AbstractPage;
 export 'src/pages/material.dart'                              show AbstractMaterialPage;
 export 'src/pages/bloc_material.dart'                         show BlocMaterialPage;
 export 'src/pages/stateless_material.dart'                    show StatelessMaterialPage;
-
-export 'src/route_information_parsers/page_stack.dart'        show PageStackRouteInformationParser;
 
 export 'src/widgets/navigator.dart'                           show PageStackBlocNavigator;