+++ lib/src/blocs/page/event.dart
@@ -2,4 +2,6 @@
 ///
 /// Extend this for your custom events like emitting notifications
 /// to be shown in overlays etc.
-abstract class PageBlocEvent {}
+abstract class PageBlocEvent {
+  const PageBlocEvent();
+}