+++ lib/src/blocs/page_stack/event.dart
@@ -1,1 +1,4 @@
-abstract class PageStackBlocEvent {}
+/// The base class for [PageStackBloc] events.
+abstract class PageStackBlocEvent {
+  const PageStackBlocEvent();
+}