+++ lib/src/blocs/page_stacks/current_page_stack_changed_event.dart
@@ -1,0 +1,11 @@
+import 'event.dart';
+
+class CurrentPageStackChangedEvent extends PageStacksBlocEvent {
+  final String? oldKey;
+  final String? newKey;
+
+  const CurrentPageStackChangedEvent({
+    required this.oldKey,
+    required this.newKey,
+  });
+}