+++ lib/src/blocs/page_stacks/event.dart
@@ -1,0 +1,4 @@
+/// The base class for [PageStacksBloc] events.
+abstract class PageStacksBlocEvent {
+  const PageStacksBlocEvent();
+}